+++ packages/quantic/force-app/main/default/lwc/quanticCategoryFacet/quanticCategoryFacet.js
@@ -1,12 +1,18 @@
 import {api, LightningElement, track} from 'lwc';
 import {registerComponentForInit, initializeWithHeadless} from 'c/quanticHeadlessLoader';
-import {I18nUtils} from 'c/quanticUtils';
+import {I18nUtils, regexEncode} from 'c/quanticUtils';
 
 import clear from '@salesforce/label/c.quantic_Clear';
 import showMore from '@salesforce/label/c.quantic_ShowMore';
 import showLess from '@salesforce/label/c.quantic_ShowLess';
 import showMoreFacetValues from '@salesforce/label/c.quantic_ShowMoreFacetValues';
 import showLessFacetValues from '@salesforce/label/c.quantic_ShowLessFacetValues';
+import allCategories from '@salesforce/label/c.quantic_AllCategories';
+import search from '@salesforce/label/c.quantic_Search';
+import moreMatchesFor from '@salesforce/label/c.quantic_MoreMatchesFor';
+import noMatchesFor from '@salesforce/label/c.quantic_NoMatchesFor';
+import collapseFacet from '@salesforce/label/c.quantic_CollapseFacet';
+import expandFacet from '@salesforce/label/c.quantic_ExpandFacet';
 
 export default class QuanticCategoryFacet extends LightningElement {
   /** @type {import("coveo").CategoryFacetState} */
@@ -24,18 +30,32 @@ export default class QuanticCategoryFacet extends LightningElement {
   @api label;
   /** @type {string} */
   @api engineId;
+  /** @type {boolean} */
+  @api noSearch = false;
 
   /** @type {import("coveo").CategoryFacet}} */
   facet;
   /** @type {import("coveo").Unsubscribe} */
   unsubscribe;
-
+  /** @type {string} */
+  collapseIconName = 'utility:dash';
+  /** @type {boolean} */
+  isExpanded = true;
+  /** @type {HTMLInputElement} */
+  input;
+ 
   labels = {
     clear,
     showMore,
     showLess,
     showMoreFacetValues,
     showLessFacetValues,
+    allCategories,
+    search,
+    moreMatchesFor,
+    noMatchesFor,
+    collapseFacet,
+    expandFacet,
   }
 
   connectedCallback() {
@@ -44,6 +64,7 @@ export default class QuanticCategoryFacet extends LightningElement {
 
   renderedCallback() {
     initializeWithHeadless(this, this.engineId, this.initialize);
+    this.input = this.template.querySelector('.facet__searchbox-input');
   }
 
   disconnectedCallback() {
@@ -72,22 +93,20 @@ export default class QuanticCategoryFacet extends LightningElement {
     return this.state.values;
   }
 
-  get parents() {
-    return this.state.parents;
+  get nonActiveParents() {
+    return this.state.parents?.slice(0, -1);
+  }
+
+  get activeParent() {
+    return this.state.parents?.slice(-1)[0];
   }
 
   get canShowMore() {
-    if (!this.facet) {
-      return false;
-    }
-    return this.state.canShowMoreValues;
+    return this.facet && this.state.canShowMoreValues  && !this.isFacetSearchActive;
   }
 
   get canShowLess() {
-    if (!this.facet) {
-      return false;
-    }
-    return this.state.canShowLessValues;
+    return this.facet && this.state.canShowLessValues;
   }
 
   get hasParents() {
@@ -96,6 +115,33 @@ export default class QuanticCategoryFacet extends LightningElement {
 
   get hasValues() {
     return this.state.values.length !== 0;
+  }
+
+  get hasSearchResults() {
+    return this.facet?.state.facetSearch.values.length !== 0;
+  }
+
+  get canShowMoreSearchResults() {
+    return this.facet?.state.facetSearch.moreValuesAvailable;
+  }
+
+  get facetSearchResults() {
+    const results = this.facet.state.facetSearch.values;
+    return results.map((result, index) => ({
+      value: result.rawValue,
+      index: index,
+      numberOfResults: result.count,
+      path: result.path,
+      localizedPath: this.buildPath(result.path) ,
+      highlightedResult: this.highlightResult(
+        result.displayValue,
+        this.input.value
+      ),
+    }));
+  }
+
+  get query() {
+    return this.input?.value;
   }
 
   get hasParentsOrValues() {
@@ -110,11 +156,47 @@ export default class QuanticCategoryFacet extends LightningElement {
     return I18nUtils.format(this.labels.showLessFacetValues, this.label)
   }
 
+  get moreMatchesForLabel() {
+    return I18nUtils.format(this.labels.moreMatchesFor, this.query);
+  }
+
+  get noMatchesForLabel() {
+    return I18nUtils.format(this.labels.noMatchesFor, this.query);
+  }
+
+  get actionButtonIcon() {
+    return this.isExpanded ? 'utility:dash' : 'utility:add';
+  }
+
+  get actionButtonLabel() {
+    const label = this.isExpanded ? this.labels.collapseFacet : this.labels.expandFacet;
+    return I18nUtils.format(label, this.label);
+  }
+
+  get isSearchComplete() {
+    return !this.facet.state.isLoading;
+  }
+
+  get isFacetSearchActive() {
+    return this.input?.value !== '';
+  }
+
   /**
    * @param {CustomEvent<import("coveo").CategoryFacetValue>} evt
    */
   onSelect(evt) {
-    this.facet.toggleSelect(evt.detail);
+    const specificSearchResult = {
+      displayValue: evt.detail.value,
+      rawValue: evt.detail.value,
+      count: evt.detail.numberOfResults,
+      path: evt.detail.path
+    };
+    if (this.isFacetSearchActive) {
+      this.facet.facetSearch.select(specificSearchResult);
+    } else {
+      this.facet.toggleSelect(evt.detail);
+    }
+    this.clearInput();
   }
 
   showMore() {
@@ -127,5 +209,47 @@ export default class QuanticCategoryFacet extends LightningElement {
 
   reset() {
     this.facet.deselectAll();
+  }
+
+  toggleFacetVisibility() {
+    if (this.isExpanded) {
+      this.clearInput();
+    }
+    this.isExpanded = !this.isExpanded;
+  }
+
+  handleKeyUp() {
+    if (this.isSearchComplete) {
+      this.facet.facetSearch.updateText(this.input?.value);
+      this.facet.facetSearch.search();
+    }
+  }
+
+  clearInput() {
+    if(this.input) {
+      this.input.value = '';
+    }
+    this.facet.facetSearch.updateText('');
+  }
+
+  highlightResult(result, query) {
+    if (!query || query.trim() === '') {
+      return result;
+    }
+    const regex = new RegExp(`(${regexEncode(query)})`, 'i');
+    return result.replace(regex, '<b>$1</b>');
+  }
+
+  /**
+  * @param {string[]} path
+  */
+  buildPath(path) {
+    if(!path.length) {
+      return this.labels.allCategories;
+    } 
+    if(path.length > 2) {
+      path = path.slice(0, 1).concat("...", ...path.slice(-1));
+    }
+    return path.join('/');
   }
 }